Create README.md
VolunShip is an Android mobile app that optimizes the process for users to volunteer from home. The app is divided into four distinct sections: the home page, instructions page, charities page, and a share page. A navigation bar was implemented for the separate components with the purpose of user convenience. The home page is where users will find specific items they can donate to a nonprofit organization. When the user taps on the item, a brief description about the impact and organization will be displayed, along with a button to the shipping label. The instructions page lists out the process/steps for the user to follow to volunteer from home. The charities page organizes and lists specific charities/causes the user can donate their items. Buttons with the links are neatly displayed for the user. The share page allows the user to easily contact others about the organization they volunteered for and their impact. For convenience purposes, any link that is clicked will be automatically copied to the user's Android Clipboard.

Link to app demonstration: https://youtu.be/fQFK5gTiL84